Add new card components and refactor panel configuration: Introduce cardBase.vue, cardPrimaryButton.vue, cardTitle.vue, midiaInsideCards.vue, numberInsideCards.vue, and paragraphInsideCards.vue for enhanced card functionality. Update panel configuration files for improved section management and integrate new components into the existing structure for better user experience.
