+++ src/components/cards/components-cards/cardInfo.vue
