+++ src/components/cards/components-cards/cardsGraph/cardGraphBase.vue
