+++ src/components/cards/components-cards/cardsGraph/graphics-base/LineGraphic.vue
