+++ src/components/cards/components-cards/cardsGraph/vegetation/VGraphicData.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script>
-import CardBase from '@/components/cards-new/components-cards/cardsGraph/cardGraphBase.vue';
+import CardBase from '@/components/cards/components-cards/cardsGraph/cardGraphBase.vue';
 import MoreVegetationImage from '@/assets/images/cards/moreVegetation.png';
 import LessVegetationImage from '@/assets/images/cards/lessVegetation.png';
