+++ src/components/cards/components-cards/cardsGraph/vegetation/VGraphicGenerator.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import CardBase from '@/components/cards-new/components-cards/cardsGraph/cardGraphBase.vue';
+import CardBase from '@/components/cards/components-cards/cardsGraph/cardGraphBase.vue';
 
 export default {
   name: 'GraphicGenerator',
