+++ src/components/cards/components-cards/cardsGraph/vegetation/VGraphicSection.vue
