+++ src/components/cards/components-cards/cardsGraph/weather/TGraphicSection.vue
