+++ src/components/cards/panel-config/base/SectionCards.vue
