+++ src/components/cards/panel-config/base/YearPicker.vue
